Add support for camera movement
The camera can now move and rotate freely,
works pretty well. Needs some polishing.
When the FoV is changed from the default value,
there's a bit of jittering: the more extreme the change,
the bigger the jitter - looks like maybe the game is
perturbing the camera values (on top of the injected values).

Currently no way to pause the gameplay or freeze time.