Add Command. Better API for Toggler and Throttler
Changed the API of the Toggler and Throttler classes to be more sensible,
and more focused on a single thing.

Added the Command (abstract) base class to represent various user commands, and an implementation of it called ToggleCommand.
(Tested it by using it as means to toggle the free cam on/off - didn't check in that as part of this commit.)

Moved the conditional stuff that creapt into the Throttler there.